Refactor require/load macros
Let's reuse the comptime_string method from elsewhere!
